Setup cross platform building and running for tests
Refactored script for building and running tests. Now
`builds/scripts/run-tests.sh` contains a series of functions that can be
individually called as part of a CI pipeline, there's also a function that
just runs everything for ease of use, though it shouldn't be used in a CI
pipeline. Note that `run-tests.sh` will likely move to a CI directory at some
point
- Add bash function to download and build googletest
- Remove gtest loading from spock and summit setup scripts
- add crc setup script
- Add "googletest*" to gitignore
- Add option to run_tests.sh to download and build google test then use
  it for linking
-  Allow custom GTest location on C-3PO and OLCF machines
- `make clean` now removes googletest too
- `make clobber` now removes all directories in `cholla/bin/` that start with
  `t` in an effort to remove directories generated by system tests